Remove unused imports, move service and models to store
