+++ src/app/main-feed/main-feed.component.html
@@ -5,7 +5,6 @@
     (scrolled)="onScrollEnd()"
 >
     <div class="question" *ngFor="let question of feed$ | async; index as i">
-
         <app-question [question]="question" [bookmarks]="(bookmarks$ | async) ? (bookmarks$ | async) : []"></app-question>
     </div>
 </div>
