+++ src/app/question/question.component.html
@@ -21,8 +21,6 @@
         <span class="tag" *ngFor="let tag of question.tags"> {{tag}}</span>
     </div>
     <div *ngIf="showDetails && question.body">
-        <div class="question-body" [innerHtml]="question.body | safe">
-
-        </div>
+        <div class="question-body" [innerHtml]="question.body | safe"></div>
     </div>
 </div>
