+++ src/app/store/so-index.service.spec.ts
