+++ src/app/store/so-index.service.ts
